[PPYOLOE] fix proj_conv in ptq bug

+++ ppdet/modeling/assigners/task_aligned_assigner.py
@@ -86,7 +86,7 @@ class TaskAlignedAssigner(nn.Layer):
         # negative batch
         if num_max_boxes == 0:
             assigned_labels = paddle.full(
-                [batch_size, num_anchors], bg_index, dtype=gt_labels.dtype)
+                [batch_size, num_anchors], bg_index, dtype='int32')
             assigned_bboxes = paddle.zeros([batch_size, num_anchors, 4])
             assigned_scores = paddle.zeros(
                 [batch_size, num_anchors, num_classes])
